--- src/App.css
@@ -1,31 +1,0 @@
-.Car {
-  border: 1px solid black;
-  height: 50vh;
-  color: #000;
-  width: 80vw;
-  margin: auto;
-}
-
-.car-content {
-  display: flex;
-  justify-content: space-evenly;
-  width: 80%;
-  margin: auto;
-  height: 40vh;
-  margin-top: 2rem;
-  
-}
-
-.car-item {
-  border: 1px solid black;
-  width: 30%;
-}
-
-.carousel-control-prev, .carousel-control-next {
-  width: 10% !important;
- 
-}
-
-.carousel-control-prev-icon, .carousel-control-next-icon  {
-  background-color: green;
-}
